Added reference application for the usage of all the peripherals.
The application is controlled through a switch in the common.h file,
remove the comment field to activate it.

The application will use the ADC to track deviations on the trim-
potentiometer (manually modified by the user) to emulate a sensor
with defined deviation thresholds. The measurement is done every
second and the ADC percentage value together with the RTC time-
stamp will be stored in the EEPROM upon +/-10% deviation.

The stored data is then printed out upon every button press.

Note! Currently there is a limitation of 256 bytes in total that
can be read out from the EEPROM. This is due to usage of static
buffer in the main program. To be changed further on.